fix(data): cast clingen_validity_score to float to match inference builder
real_data_prep emitted int while the inference feature builder in variant_ensemble uses float. Aligning the dtype removes a latent train/inference skew. Value-neutral today (ClinGen validity maps to integers 0-5).

+++ src/genomic_variant_classifier/data/real_data_prep.py
@@ -1054,7 +1054,7 @@ class DataPrepPipeline:
         feats["clingen_validity_score"] = (
             df.get("clingen_validity_score", pd.Series([0] * len(df), index=df.index))
             .fillna(0)
-            .astype(int)
+            .astype(float)  # match inference builder (variant_ensemble); int truncated a future fractional score
         )
 
         # HGMD
